OSAP_web/OSAP_flask/requirements.txt
beautifulsoup4==4.11.2
dominate==2.7.0
Flask==2.2.2
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.4.0
requests==2.28.1
scipy==1.9.3
torch==1.12.0
torchvision==0.14.0
visdom==0.2.3
